Add files via upload
This is just a very basic review for the exam of "theoreticl cosmology" that I took during my master's degree. It is not an original work, but rather a simple summary of articles written by other known researchers, each of wich is cited in the bibliography.
The role of neutrinos and standard model particle physics in modifying primordial gravitational waves and their power spectrum is summarized.
Also, the damping effect of free streaming neutrinos is described and the possibility of having neutrinos as actual source of GWs (very) briefly reported.